pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cyheadgen"
authors = [
    { name = "Matthias Cremieux" },
]
version = "0.1"
urls = { Home = "https://github.com/MattCrmx/cyheadgen" }
requires-python = ">=3.8"
dependencies = [
    "ply",
]

[project.optional-dependencies]
dev = [
    "black==22.3.0",
    "pre-commit==2.18.1",
    "pytest-env"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py38"
lint.select = [
    "B",
    "D",
    "E",
    "F",
    "I",
    "N",
    "PGH",
    "UP",
    "S",
    "T20",
    "TID",
    "W",
    "RET"
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D", "R", "S"]
"src/cyheadgen/_lexer.py" = ["N816", "N802", "D415"]
"src/cyheadgen/_parser.py" = ["N816", "N802", "D415", "D205", "E501"]


[tool.mypy]
files = "src"
python_version = "3.8"
ignore_missing_imports = true
strict = true

[[tool.mypy.overrides]]
module = "ply.*"
follow_imports = "skip"
